add refresh function to auth.ts to keep to DRY

+++ fileapp/src/stores/auth.ts
@@ -2,6 +2,8 @@ import { get, writable } from 'svelte/store';
 import { getCookie, setCookie, deleteCookie } from '../utils/cookies';
 import { User } from '$lib/models';
 import { goto } from '$app/navigation';
+import { API_BASE_URL } from '$lib/config';
+import { redirect } from '@sveltejs/kit';
 
 export const isLoggedIn = writable(false);
 export const useremail = writable('');
@@ -19,10 +21,32 @@ export async function Login(tokenResponse: Response, userObj: User) {
 
 export async function Logout() {
     deleteCookie("token");
-    deleteCookie("refresh")
-    deleteCookie("userinfo")
+    deleteCookie("refresh");
+    deleteCookie("userinfo");
     useremail.set('');
     isLoggedIn.set(false);
     goto("/login")
 }
 
+
+export async function Refresh() {
+    let refresh = getCookie("refresh");
+    let refreshRespone = await fetch(API_BASE_URL + "/refresh", {
+        method: 'POST',
+        headers: {
+            'content-type': 'application/json'
+        },
+        body: JSON.stringify({ refreshtoken: refresh })
+    });
+
+    if (!refreshRespone.ok) {
+        deleteCookie("userinfo");
+        deleteCookie("token");
+        deleteCookie("refresh");
+        redirect(303, "/login")
+    }
+
+    let refreshJson = await refreshRespone.json();
+    setCookie("token", refreshJson.accessToken);
+    setCookie("refresh", refreshJson.refreshToken);
+}
